Add standalone language select component
Add standalone language select component. Header and ion-content need
to be in a single component, otherwise the detail view is not shown correctly.
The component/module also needs to be located in a different directory than
the calling module and also not in a subdirectory, otherwise the defined routes
will not (at least they did not work while trying this ...)